myshop/requirements.txt
amqp==2.6.1
asgiref==3.6.0
Babel==2.8.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
bs4==0.0.1
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.7
certifi==2017.7.27.1
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
cssselect2==0.3.0
defusedxml==0.6.0
Django==3.2
django-admin-thumbnails==0.2.6
django-ckeditor==6.5.1
django-countries==6.1.3
django-crispy-forms==1.9.2
django-currencies==0.10.1
django-extensions==2.2.5
django-js-asset==2.0.0
django-phonenumber-field==7.0.2
django-taggit==1.2.0
django-tawkto==0.2
flower==0.9.5
future==0.18.2
html5lib==1.1
humanize==2.6.0
idna==2.7
jsonfield==2.1.1
kombu==4.6.11
oauthlib==3.1.0
openexchangerates==0.1.1
Paperboy==1.0.1
phonenumbers==8.13.7
Pillow==7.2.0
prometheus-client==0.8.0
psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.0.0
Pyphen==0.9.5
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2017.2
redis==3.4.1
requests==2.20.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
soupsieve==2.0.1
sqlparse==0.3.1
tinycss2==1.0.2
tornado==5.1.1
urllib3==1.24.3
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
Werkzeug==0.16.0
